Move order page lambda queries to a shared file
