+++ apps/app/src/queries/index.ts
@@ -1,0 +1,1 @@
+export * from './lambda-api';
