+++ apps/app/src/queries/lambda-api/index.ts
@@ -1,0 +1,2 @@
+export * from './orders';
+export * from './pharmacies';
